Correct max_gas warning logic
My integer division trick here was boldly incorrect, the result of the
integer divsion will be one if the value is more than half and zero only
if gas_used is greater, which we already check for in the above
condition.

Therefore m can only be values of one and greater than one, where one
indicates that more than half of hte maximum gas has been used by a
single transaction.